requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
antiorm==1.2.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
db==0.1.1
db-sqlite3==0.0.1
Django==5.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
instaloader==4.10
magic-filter==1.0.11
multidict==6.0.4
pillow==10.2.0
psycopg2==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
